pyproject.toml
[project]
name = "wcm-facerec"
version = "0.1.0"
description = "Large Scale Face Recognition with DeepFace + pgvector"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "deepface>=0.0.89",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
    "pgvector>=0.3.1",
    "numpy>=2.2.0",
    "pillow>=11.1.0",
    "httpx>=0.28.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "opencv-python-headless>=4.10.0.20241230",
    "numpy<2.2.1",
    "tf-keras>=2.21.0",
    "xlrd>=2.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/wcm_facerec"]

[tool.hatch.build.targets.wheel.sources]
"wcm_facerec" = "src/wcm_facerec"
